refactor: lib/constants/error to lib/error
