+++ src/lib/events/names.ts
@@ -1,0 +1,3 @@
+export const IMPRESSION_EVENT = "impression";
+export const ADD_TO_CART_EVENT = "add_to_cart";
+export const PURCHASE_EVENT = "purchase";
